Add specification of min-sdk-version during build
The min-sdk-version for the time zone data app needs to be
specified during the build to prevent the latest .apk being
installed on *all* devices.

For example, this is what will prevent R .apks being being
installed on P devices. The R .apk will have a min-sdk-version
matching the R API number, while the P .apk will have one
matching the P API number. When a P device is looking for
the a matching .apk it won't be able to use the R one (since
that .apk requires R min), so the device will pick the latest
P one instead. This relies on the application version codes
being arranged sensibly too, as per
https://developer.android.com/google/play/publishing/multiple-apks.html

Bug: 31008728
Test: Manual testing
Change-Id: Idd6d06271e035e9d6f8900b4a859cf4bb0c04570